SCUMM: Fix broken text speed workaround
The value is saved to the Config Manager but _defaultTalkDelay is only updated the next time syncSoundSettings() is called, or the next time the user changes the text speed, which is based on _defaultTalkDelay, effectively ignoring the value set here.

Also, overriding VAR_CHARINC here breaks the script-based text speed controls.