AGI: Avoid mapping mystery keys to F11 and F12
This is a remnant of the Sarien project. These macros were later renamed F11 and F12 in ScummVM, however these magic numbers are not derived from the F11 and F12 keys, nor from any other key in DOS. A Google search does not shed light on the original macro names.