SCUMM: Enable Snap Scroll more carefully
In MANIAC V1, in the kitchen, with Snap Scroll enabled, you may warp the screen past the trigger for Edna to start chasing you. She will remain immobile and only reacts once you walk on top of her.

To categorically get rid of these types of bug, it would be best to enable Snap Scroll only where it exists in the original games, i.e. disable it for V0, V1 and V5.

In V5, the scripts manipulate VAR_CAMERA_FAST so it is a bit scary if we mess with that variable unexpectedly.