SCI: Remove extraneous LSL3 submarine control room workaround
The problem here was slightly misinterpreted. There are two things happening: First, when you enter the submarine control room, the game speed selector is set to 1, the fastest speed. This is slightly peculiar, but the player can still change the game speed manually. However, in this room the + and - keys are actually part of the submarine control scheme, while using + or - to change the game speed is _not_ disabled.

This is already bugged, but worse; when you press -, which you would expect to decrease the game speed (increasing the game speed selector), the game speed selector is surprisingly set 0, a speed otherwise unattainable in SCI0.

In my opinion, the expected behavior is that the keyboard speed controls are disabled in this room. Nothing else is broken, and so we do not need this workaround.