SKY: Fix frame-time to millisecond conversion
It seems that, in some refactoring of this code, something got mangled with the conversion factors because the default game speed (converted) does not convert cleanly to the relative speed slider position i.e. it is not a multiple of 12. The default speed is correct, all speeds produced by the speed slider are completely wrong.

I would say I am surprised that this was here for all these years.