SCUMM: Clean up dissolveEffect() slightly
It is best if the blits_before_refresh divisor is a common factor of all possible values of w * h.

There is no reason to sleep for longer than a frame for each call to waitForTimer(). Instead it would be natural to use VAD_FADE_DELAY here like in the other screen effects (even though the delay is completely arbitrary).