SCI: Update the screen unconditionally in SciEngine::sleep()
In kWait loops slower than ticks == 1, the screen update throttling will interfere with the expected screen updates in SciEngine::sleep(). Up to every other screen update call will be skipped, which causes noticeable cursor lag.

This is a fundamental problem with this implementation. We cannot throttle the screen updates whilst sleeping with less granularity than will trigger them consistently.

We need to update the screen unconditionally with g_system->updateScreen(), or maximize the sleep granularity, with delayMillis(1), to make sure we get a screen update when it is available.