SCI: Amend KQ6 patch for talking inventory items
I think that sluicebox deliberately did not use kWait in the original patch, however now that the speedthrottler can be disabled by the user, it would be best not to rely on it in the script patches. Using kWait would get rid of the ugly source code part of this patch, as well.