SCUMM: Adjust shake position offsets
In common/system.h, it says that "The effect causes the displayed graphics to be shifted upwards by the specified (always positive) offset.", but that is not true; a positive offset produces a downward shift, and so we need negative offsets in SCUMM.

The old offsets were doubled. I guess this was an adjustment for something, in a time gone by. Finally, the order of the offsets, i.e. the starting position, was incorrect.