SCI: Implement simulated VGA vertical retrace synchronization
The original interpreter reads a VGA status register (port 3DA), waiting for the vertical retrace bit to be set, which synchronizes the animation steps to the refresh rate for these various screen animations.

I do not fully understand the significance of this but at least in SQ6, for kSetFade, the screen transition duration remains constant between transitions with one or multiple steps per intensity level. There was a patch for an extreme case in SQ6 but this happens all throughout that game; perhaps in other games as well although I did not find any non-SQ6 cases yet. We should only sleep once per intensity level essentially.