SCI: Normalize SCI16 screen transition delays
In GfxTransitions::straight(), we skip every other frame to account for the fewer number of steps in that direction, but that was already accounted for in the delay increment. Now the vertical transition is relatively over two times slower than it should be.

The screen transition delays are a bit all over the place anyway. I know the delays are completely arbitrary, but the relative duration of each screen transitions should be roughly equivalent.

In general, the screen transitions in SCI0 are not an amazing feature, and you will almost certainly not have seen them unless playing ScummVM (they are virtually instantaneous on an ao486).

Locking up the game for a second on every screen gets old, and it was not their intended purpose. I still left them at a very slow speed, but disabling the speedthrottler will disable the delay.